ParserImpl free of mutable state, except for Scanner
Scanner could be threaded through all calls as a parameter instead of
being part of the state of the Parser, but I'm not seeing what
advantage that would bring: Parsing is inherently stateful at some level
since the source of characters is stateful in the sense that we want to
read each character exactly once and in the correct order.

Currently parsing is implemented in three layers:

- Parser
- Scanner
- Readable

Conceptually, we can divide it even more finely:

5. Parser (source of Values)
4. Scanner (source of Tokens)
3. Buffered Source of Characters granting lookahead of 1
2. Source of Characters
1. Readable

(1) is inherently stateful. (2) is inherently stateful, transitively.
(3) is inherently stateful because the last character looked at by
    one call to the scanner, might need to be looked at first by
    the next call to the scanner.
(4) needs (3), so 4 is stateful.
(5) is stateful because it uses (4).

Now, we could break this transitive chain of statefulness at any point,
by incorporating more input parameters into the methods to be called
and/or returning more complex return values. The scanner could return
the recognized token, and the state necessary for the next call. But
this is no solution. This doesn't solve the state problem, it just makes
it someone else's problem.

The fact is there's nothing to be gained by making the scanner or the
parser "thread-safe". They are performing an inherently serial task
across a single source of input. They are designed to be light-weight
single-use objects, much like the Readable they are ultimately built
upon. If you want to parse from multiple input streams concurrently,
use multiple Parsers. It's not like they are expensive to create or
to dispose of.