remove mutable state from Parser+Scanner
This implies that Parser.nextValue(…) now takes a PushbackReader, which
it necessarily mutates, as an argument.

The upside is that Parser (and Scanner) are now immutable and thread-
safe.

The downside is that the user of a Parser now must keep track of two
objects: the Parser and the PushbackReader.

Internally, Scanner and Parser appear more complex because they are
threading up the method call chain all the state that they used to
simply contain.

I'm not yet convinced that this is such a win.