Provide extremely simple unmarshal of untagged maps to java objects
An Unmarshaller will walk a tree of EDN values, calling transformation
functions on sub-maps. Transformation functions are keyed by the map
key under which the Map they are called upon is in its parent map.
These keys must be keywords. No context is taken into account: any
given key always means the same thing to a given Unmarshaller,
irrespective of what kind of map it may appear in.

Despite these limitations, it allows us to turn this:

    {:upperLeft {:x 1 :y 2}, :lowerRight {:x 3, :y 4}}

Into this:

    new Rect(new Point(1, 2), new Point(3, 4));

Without needing to do the traversal ourselves, provided we configure
the Unmarshaller appropriately, and know that the top-level map should
be treated as a rectangle.

See UnmarshallersTest for an example.